refactor(renderer, materials): move files
