+++ packages/materials/vue-opentiny-vue/components/i18n.ts
@@ -1,0 +1,1 @@
+// TODO: setup i18n here
